// src/rides/enums/rides.enum.ts
export enum RIDE_STATUS {
  PENDING = "PENDING",
  DISPATCHED = "DISPATCHED",
  HELD = "HELD",
  RELEASED = "RELEASED",
  RECOVER = "RECOVER",
  ACCEPTED = "ACCEPTED",
  REJECTED = "REJECTED",
  ARRIVED = "ARRIVED",
  PICKEDUP = "PICKEDUP",
  DROPOFF = "DROPOFF",
  CANCELLED = "CANCELLED",
  NO_FARE = "NO_FARE",
  COMPLETED = "COMPLETED",
  REQUESTED = "REQUESTED",
  FOB_REQUESTED = "FOB_REQUESTED",
  FOB_DISPATCHED = "FOB_DISPATCHED",
  FOB_ACCEPTED = "FOB_ACCEPTED",
  FOB_REJECTED = "FOB_REJECTED",
  OVERRIDE_REQUESTED = "OVERRIDE_REQUESTED",
  OVERRIDE_DISPATCHED = "OVERRIDE_DISPATCHED",
  OVERRIDE_REJECTED = "OVERRIDE_REJECTED",
  FOB_OVERRIDE_REQUESTED = "FOB_OVERRIDE_REQUESTED",
  FOB_OVERRIDE_DISPATCHED = "FOB_OVERRIDE_DISPATCHED",
  FOB_OVERRIDE_REJECTED = "FOB_OVERRIDE_REJECTED",
}

//pending held and recovered show krwani h

export enum RIDE_PAYMENT_STATUS {
  UNPAID = "UNPAID",
  PAID = "PAID",
}

export enum ERIDE_PAYMENT_METHOD {
  ONLINE = "ONLINE",
  CASH = "CASH",
  CARD = "CARD",
  GOOGLE_PAY = "GOOGLE_PAY",
  APPLE_PAY = "APPLE_PAY",
}

export enum RIDE_HISTORY_ACTION_TYPE {
  CREATED = "CREATED",
  MODIFIED = "MODIFIED",
  DELETED = "DELETED",
}

export enum STATUS {
  ACTIVE = "ACTIVE",
  DISABLED = "DISABLED",
  SUSPENDED = "SUSPENDED",
}